simplify ENT system #56
