--- pronotepy/ent.py
@@ -1,594 +1,0 @@
-# type: ignore
-from logging import getLogger, DEBUG
-import typing
-
-import requests
-from bs4 import BeautifulSoup
-from urllib.parse import urlparse, parse_qs
-
-from .exceptions import *
-
-log = getLogger(__name__)
-log.setLevel(DEBUG)
-
-HEADERS = {
-    'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:73.0) Gecko/20100101 Firefox/73.0'
-}
-
-
-@typing.no_type_check
-def educonnect(url: str, session, username: str, password: str):
-    payload = {
-    'j_username': username,
-    'j_password': password,
-    '_eventId_proceed': ''
-    }
-    response = session.post(url, headers=HEADERS, data=payload)
-    # 2nd SAML Authentication
-    soup = BeautifulSoup(response.text, 'html.parser')
-    input_SAMLResponse = soup.find("input", {"name": "SAMLResponse"})
-    if not input_SAMLResponse:
-        return
-
-    payload = {
-        "SAMLResponse": input_SAMLResponse["value"]
-        }
-
-    input_relayState = soup.find("input", {"name": "RelayState"})
-    if input_relayState:
-        payload["RelayState"] = input_relayState['value']
-
-    response = session.post(soup.find("form")["action"], headers=HEADERS, data=payload)
-    return response
-
-
-def ac_grenoble(username: str, password: str):
-    """
-    ENT ac Grenoble
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    identity_provider = "https://cas.ent.auvergnerhonealpes.fr/login?selection=EDU"
-
-    # SAML authentication
-    session = requests.Session()
-    response = session.get(identity_provider, headers=HEADERS)
-
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"]
-        }
-
-    log.debug(f'[ENT Eaux claires] Logging in with {username}')
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def ac_rennes(username: str, password: str):
-    """
-    ENT ac Rennes Toutatice.fr
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # Toutatice required URL
-    toutatice_url = 'https://www.toutatice.fr/portail/auth/MonEspace'
-    toutatice_login = 'https://www.toutatice.fr/wayf/Ctrl'
-    toutatice_auth = 'https://www.toutatice.fr/idp/Authn/RemoteUser'
-
-    session = requests.Session()
-
-    response = session.get(toutatice_url, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "entityID": soup.find("input", {"name": "entityID"})["value"],
-        "return": soup.find("input", {"name": "return"})["value"],
-        "_saml_idp": soup.find("input", {"name": "_saml_idp"})["value"]
-        }
-
-    log.debug(f'[ENT Toutatice] Logging in with {username}')
-    response = session.post(toutatice_login, data=payload, headers=HEADERS)
-
-    educonnect(response.url, session, username, password)
-
-    params = {
-        'conversation': parse_qs(urlparse(response.url).query)['execution'][0],
-        'redirectToLoaderRemoteUser': 0,
-        'sessionid': session.cookies.get('IDP_JSESSIONID')
-        }
-
-    response = session.get(toutatice_auth, headers=HEADERS, params=params)
-    soup = BeautifulSoup(response.text, 'xml')
-
-    if soup.find('erreurFonctionnelle'):
-        raise(PronoteAPIError('Toutatice ENT (ac_rennes) : ', soup.find('erreurFonctionnelle').text))
-    elif soup.find('erreurTechnique'):
-        raise(PronoteAPIError('Toutatice ENT (ac_rennes) : ', soup.find('erreurTechnique').text))
-    else:
-        params = {
-            'conversation': soup.find("conversation").text,
-            'uidInSession': soup.find("uidInSession").text,
-            'sessionid': session.cookies.get('IDP_JSESSIONID')
-            }
-        t = session.get(toutatice_auth, headers=HEADERS, params=params)
-
-    return session.cookies
-
-
-def atrium_sud(username: str, password: str):
-    """
-    ENT for Atrium Sud
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://www.atrium-sud.fr/connexion/login'
-
-    # ENT Connection
-    session = requests.Session()
-    response = session.get(ent_login, headers=HEADERS)
-
-    log.debug(f'[ENT Atrium] Logging in with {username}')
-
-    # Login payload
-    soup = BeautifulSoup(response.text, 'html.parser')
-    input_ = soup.find('input', {'type': 'hidden', 'name': 'execution'})
-    execution = input_.get('value')
-
-    payload = {
-        'execution': execution,
-        '_eventId': 'submit',
-        'submit': '',
-        'username': username,
-        'password': password
-        }
-
-    # Send user:pass to the ENT
-    cookies = session.cookies
-    response = session.post(ent_login, headers=HEADERS, data=payload, cookies=cookies)
-    return session.cookies
-
-
-def ac_reims(username: str, password:str):
-    """
-    ENT for AC Reims
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://services-familles.ac-reims.fr/login/ct_logon_vk.jsp?CT_ORIG_URL=%2Fsso%2FSSO%3FSPEntityID%3DSP-MonBureauNumerique-Production&ct_orig_uri=%2Fsso%2FSSO%3FSPEntityID%3DSP-MonBureauNumerique-Production'
-    ent_verif = 'https://services-familles.ac-reims.fr/aten-web/connexion/controlesConnexion?CT_ORIG_URL=%2Fsso%2FSSO%3FSPEntityID%3DSP-MonBureauNumerique-Production&ct_orig_uri=%2Fsso%2FSSO%3FSPEntityID%3DSP-MonBureauNumerique-Production'
-    pronote_verif = 'https://cas.monbureaunumerique.fr/saml/SAMLAssertionConsumer'
-    # ENT Connection
-    session = requests.Session()
-    response = session.get(ent_login, headers=HEADERS)
-    log.debug(f'[ENT AC Reims] Logging in with {username}')
-    # Send user:pass to the ENT
-    cookies = session.cookies
-    # Login payload
-    payload = {
-        'auth_mode': 'BASIC',
-        'orig_url': '/sso/SSO?SPEntityID=SP-MonBureauNumerique-Production',
-        'user': username,
-        'password': password
-        }
-    response = session.post(ent_login, headers=HEADERS, data=payload, cookies=cookies)
-    # Get the CAS verification shit
-    cookies = session.cookies
-    response = session.get(ent_verif, headers=HEADERS, cookies=cookies)
-    # Get the actual values
-    soup = BeautifulSoup(response.text, 'html.parser')
-    cas_infos = {}
-    inputs = soup.findAll('input', {'type': 'hidden'})
-    for input_ in inputs:
-        cas_infos[input_.get('name')] = input_.get('value')
-    cookies = session.cookies
-    session.cookies.update({'SERVERID': 'gdest-prod-web14', 'preselection': 'REIMS-ATS_parent_eleve'})
-    response = session.post(pronote_verif, headers=HEADERS, data=cas_infos, cookies=cookies)
-    # Get Pronote
-    cookies = session.cookies
-    return cookies
-
-
-def occitanie_montpellier(username: str, password: str):
-    """
-    ENT for Occitanie Montpellier
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://cas.mon-ent-occitanie.fr/login'
-    # ENT Connection
-    session = requests.Session()
-
-    params = {
-        'selection': 'MONT-EDU_parent_eleve',
-        'submit': 'Valider'
-    }
-
-    response = session.get(ent_login, params=params, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"]
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    log.debug(f'[ENT Occitanie] Logging in with {username}')
-
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def occitanie_toulouse(username: str, password: str):
-    """
-    ENT for Occitanie Toulouse
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://cas.mon-ent-occitanie.fr/login'
-    # ENT Connection
-    session = requests.Session()
-
-    params = {
-        'selection': 'TOULO-ENT_parent_eleve',
-        'submit': 'Valider'
-    }
-
-    response = session.get(ent_login, params=params, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {}
-    for input_ in soup.find('form', class_='cas__login-form').findAll('input'):
-        payload[input_['name']] = input_.get('value')
-    payload['username'] = username
-    payload['password'] = password
-
-    session.post(ent_login, data=payload, headers=HEADERS)
-
-    return session.cookies
-
-
-def ac_reunion(username: str, password: str):
-    """
-    ENT for AC Reunion
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://portail.college-jeandesme.re:8443/login?service=https:%2F%2Fportail.college-jeandesme.re%2Fpronote%2Feleve.html'
-
-    # ENT Connection
-    session = requests.Session()
-    response = session.get(ent_login, headers=HEADERS)
-
-    log.debug(f'[ENT Reunion] Logging in with {username}')
-
-    # Login payload
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        'service': 'https://portail.college-jeandesme.re/pronote/eleve.html',
-        'lt': soup.find('input', {'type': 'hidden', 'name': 'lt'}).get('value'),
-        'previous_user': f'{username}@default',
-        'username': username,
-        'password': password,
-    }
-    # Send user:pass to the ENT
-    cookies = session.cookies
-    response = session.post(ent_login, headers=HEADERS, data=payload, cookies=cookies)
-
-    new_url = response.url
-
-    cookies = session.cookies
-    response = session.get(new_url, headers=HEADERS, cookies=cookies)
-
-    pronote_url = response.url
-
-    cookies = session.cookies
-    response = session.get(pronote_url, headers=HEADERS, cookies=cookies)
-
-    return session.cookies
-
-
-def ac_lyon(username, password):
-    """
-    ENT for Lyon
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://cas.ent.auvergnerhonealpes.fr/login?selection=LYON-ATS_parent_eleve&submit=Valider'
-
-    # ENT Connection
-    session = requests.Session()
-    response = session.get(ent_login, headers=HEADERS)
-
-    soup = BeautifulSoup(response.text, 'html.parser')
-
-    cookies = session.cookies
-
-    payload = {
-        'username': username,
-        'password': password,
-        'selection': "LYON-ATS_parent_eleve",
-        'codeFournisseurIdentite': "ATS-LYON",
-        '_eventId': "submit",
-        'submit': "Confirm",
-        'geolocation': "",
-        'execution': soup.find('input', {'type': 'hidden', 'name': 'execution'}).get('value')
-        }
-
-    response = session.post(ent_login, headers=HEADERS, data=payload, cookies=cookies)
-
-    return session.cookies
-
-
-def ac_orleans_tours(username, password):
-    """
-    ENT for AC Orleans-Tours
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://ent.netocentre.fr/cas/login?&idpId=parentEleveEN-IdP'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return requests.utils.cookiejar_from_dict(
-        requests.utils.dict_from_cookiejar(session.cookies))
-
-
-def monbureaunumerique(username, password):
-    """
-    ENT for MonBureauNumerique (Grand Est)
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://cas.monbureaunumerique.fr/login?selection=EDU'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"],
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return requests.utils.cookiejar_from_dict(
-        requests.utils.dict_from_cookiejar(session.cookies))
-
-
-def ent_elyco(username, password):
-    ent_login_page = 'https://cas3.e-lyco.fr/discovery/WAYF'
-    session = requests.Session()
-
-    params = {
-        'entityID': 'https://cas3.e-lyco.fr/shibboleth',
-        'returnX': 'https://cas3.e-lyco.fr/Shibboleth.sso/Login',
-        'returnIDParam': 'entityID',
-        'action': 'selection',
-        'origin': 'https://educonnect.education.gouv.fr/idp'
-    }
-
-    response = session.get(ent_login_page, params=params, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def open_ent_ng(url: str, username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for given url. If a specific function for your ENT already exists you should use it!
-
-    Parameters
-    ----------
-    url : str
-        url of the ENT
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT Connection
-    session = requests.Session()
-
-    payload = {
-        'email': username,
-        'password': password
-    }
-    response = session.post(url, headers=HEADERS, data=payload)
-    return session.cookies
-
-
-def l_normandie(username: str, password: str):
-    """
-    ENT l'educ de Normandie
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-
-    # URL required
-    ent_login_page = 'https://educonnect.education.gouv.fr/idp/profile/SAML2/Unsolicited/SSO'
-
-    session = requests.Session()
-
-    params = {
-        'providerId': 'https://ent.l-educdenormandie.fr/auth/saml/metadata/idp.xml'
-    }
-
-    response = session.get(ent_login_page, params=params, headers=HEADERS)
-    response = educonnect(response.url, session, username, password)
-
-    if not response:
-        return open_ent_ng(response.url, username, password)
-    else:
-        soup = BeautifulSoup(response.text, 'html.parser')
-        if soup.find('title').get_text() == 'Authentification':
-            return open_ent_ng(response.url, username, password)
-
-    log.debug(f'[ENT Educ de Normandie] Logging in with {username}')
-
-    return session.cookies
-
-
-def ent_essonne(username: str, password: str):
-    return open_ent_ng('https://www.moncollege-ent.essonne.fr/auth/login', username, password)
-
-
-def ent_hdf(username: str, password: str):
-    return open_ent_ng('https://enthdf.fr/auth/login', username, password)
-
-
-def ile_de_france(username: str, password: str):
-    return open_ent_ng('https://ent.iledefrance.fr/auth/login', username, password)
-
-
-def paris_classe_numerique(username: str, password: str):
-    return open_ent_ng('https://ent.parisclassenumerique.fr/auth/login', username, password)
-
-
-def ent77(username: str, password: str):
-    return open_ent_ng('https://ent77.seine-et-marne.fr/auth/login', username, password)
-
-
-def ent_mayotte(username: str, password: str):
-    return open_ent_ng('https://mayotte.opendigitaleducation.com/auth/login', username, password)
-
-
-def lyceeconnecte_aquitaine(username: str, password: str):
-    return open_ent_ng('https://mon.lyceeconnecte.fr/auth/login#/', username, password)
-
-
-def neoconnect_guadeloupe(username: str, password: str):
-    return open_ent_ng('https://neoconnect.opendigitaleducation.com/auth/login', username, password)
