+++ pronotepy/ent/__init__.py
@@ -1,0 +1,4 @@
+from .simple_auth import *
+from .educonnect import *
+from .open_ent_ng import *
+from .open_ent_ng_educonnect import *
